Advanced training regimes (#71)
Resolves #30, #33, #36, #37, #38, #42, #43, #47, #66, #80, #81

+++ aligner/models.py
@@ -22,6 +22,8 @@ class Archive(object):
     (https://github.com/prosodylab/Prosodylab-Aligner) archive class.
     """
 
+    extension = '.zip'
+
     def __init__(self, source, is_tmpdir=False):
         self._meta = {}
         self.name, _ = os.path.splitext(os.path.basename(source))
@@ -81,67 +83,63 @@ class AcousticModel(Archive):
             yaml.dump(aligner.meta, f)
 
     @property
+    def feature_config(self):
+        from .features.config import FeatureConfig
+        fc = FeatureConfig(self.dirname)
+        fc.update(self._meta['features'])
+        return fc
+
+    @property
     def meta(self):
+        default_features = {'type': 'mfcc',
+                            'use_energy': False,
+                            'frame_shift': 10,
+                            'pitch': False}
         if not self._meta:
             meta_path = os.path.join(self.dirname, 'meta.yaml')
             if not os.path.exists(meta_path):
                 self._meta = {'version': '0.9.0',
-                              'architecture': 'gmm-hmm'}
+                              'architecture': 'gmm-hmm',
+                              'features': default_features
+                              }
             else:
                 with open(meta_path, 'r', encoding='utf8') as f:
                     self._meta = yaml.load(f)
+                if self._meta['features'] == 'mfcc+deltas':
+                    self._meta['features'] = default_features
+            if 'uses_lda' not in self._meta: # Backwards compat
+                self._meta['uses_lda'] = False
+            if 'uses_sat' not in self._meta:
+                self._meta['uses_sat'] = False
+            if 'phone_type' not in self._meta:
+                self._meta['phone_type'] = 'triphone'
             self._meta['phones'] = set(self._meta.get('phones', []))
         return self._meta
 
-    def add_triphone_model(self, source):
+    def add_lda_matrix(self, source):
+        copyfile(os.path.join(source, 'lda.mat'), os.path.join(self.dirname, 'lda.mat'))
+
+    def add_ivector_model(self, source):
+        copyfile(os.path.join(source, 'final.ie'), os.path.join(self.dirname, 'final.ie'))
+        copyfile(os.path.join(source, 'final.dubm'), os.path.join(self.dirname, 'final.dubm'))
+
+    def add_model(self, source):
         """
         Add file into archive
         """
-        copyfile(os.path.join(source, 'final.mdl'), os.path.join(self.dirname, 'ali-final.mdl'))
-        copyfile(os.path.join(source, 'final.occs'), os.path.join(self.dirname, 'ali-final.occs'))
-        copyfile(os.path.join(source, 'tree'), os.path.join(self.dirname, 'ali-tree'))
+        copyfile(os.path.join(source, 'final.mdl'), os.path.join(self.dirname, 'final.mdl'))
+        if os.path.exists(os.path.join(source, 'final.occs')):
+            copyfile(os.path.join(source, 'final.occs'), os.path.join(self.dirname, 'final.occs'))
+        copyfile(os.path.join(source, 'tree'), os.path.join(self.dirname, 'tree'))
 
-    def add_triphone_fmllr_model(self, source):
-        """
-        Add file into archive
-        """
-        copy(os.path.join(source, 'final.mdl'), self.dirname)
-        copy(os.path.join(source, 'final.occs'), self.dirname)
-        copy(os.path.join(source, 'tree'), self.dirname)
-
-    def add_nnet_model(self, source):
-        """
-        Add file into archive
-        """
-        copy(os.path.join(source, 'final.mdl'), self.dirname)
-        copy(os.path.join(source, 'tree'), self.dirname)
-        for file in glob.glob(os.path.join(source, 'alignfeats.*')):
-            copy(os.path.join(source, file), self.dirname)
-        for file in glob.glob(os.path.join(source, 'fsts.*')):
-            copy(os.path.join(source, file), self.dirname)
-
-    def export_triphone_model(self, destination):
+    def export_model(self, destination):
         """
         """
         os.makedirs(destination, exist_ok=True)
         copyfile(os.path.join(self.dirname, 'final.mdl'), os.path.join(destination, 'final.mdl'))
-        copyfile(os.path.join(self.dirname, 'final.occs'), os.path.join(destination, 'final.occs'))
+        if os.path.exists(os.path.join(self.dirname, 'final.occs')):
+            copyfile(os.path.join(self.dirname, 'final.occs'), os.path.join(destination, 'final.occs'))
         copyfile(os.path.join(self.dirname, 'tree'), os.path.join(destination, 'tree'))
-
-    def export_triphone_fmllr_model(self, destination):
-        """
-        """
-        os.makedirs(destination, exist_ok=True)
-        copy(os.path.join(self.dirname, 'final.mdl'), destination)
-        copy(os.path.join(self.dirname, 'final.occs'), destination)
-        copy(os.path.join(self.dirname, 'tree'), destination)
-
-    def export_nnet_model(self, destination):
-        os.makedirs(destination, exist_ok=True)
-        copy(os.path.join(self.dirname, 'final.mdl'), destination)
-        copy(os.path.join(self.dirname, 'tree'), destination)
-        for file in glob.glob(os.path.join(self.dirname, 'fsts.*')):
-            copy(os.path.join(self.dirname, file), destination)
 
     def validate(self, dictionary):
         if isinstance(dictionary, G2PModel):
@@ -199,9 +197,58 @@ class IvectorExtractor(Archive):
     '''
     Archive for i-vector extractors (used with DNNs)
     '''
-    def export_ivector_extractor(self, destination):
+    def add_meta_file(self, trainer):
+        with open(os.path.join(self.dirname, 'meta.yaml'), 'w', encoding='utf8') as f:
+            yaml.dump(trainer.meta, f)
+
+    @property
+    def meta(self):
+        if not self._meta:
+            meta_path = os.path.join(self.dirname, 'meta.yaml')
+            with open(meta_path, 'r', encoding='utf8') as f:
+                self._meta = yaml.load(f)
+        return self._meta
+
+    def add_model(self, source):
+        """
+        Add file into archive
+        """
+        copyfile(os.path.join(source, 'final.ie'), os.path.join(self.dirname, 'final.ie'))
+        copyfile(os.path.join(source, 'final.dubm'), os.path.join(self.dirname, 'final.dubm'))
+        copyfile(os.path.join(source, 'lda.mat'), os.path.join(self.dirname, 'lda.mat'))
+
+    def export_model(self, destination):
         os.makedirs(destination, exist_ok=True)
         copy(os.path.join(self.dirname, 'final.ie'), destination)           # i-vector extractor itself
-        copy(os.path.join(self.dirname, 'global_cmvn.stats'), destination)  # Stats from diag UBM
         copy(os.path.join(self.dirname, 'final.dubm'), destination)         # Diag UBM itself
-        copy(os.path.join(self.dirname, 'final.mat'), destination)          # LDA matrix
+        copy(os.path.join(self.dirname, 'lda.mat'), destination)          # LDA matrix
+
+
+        # Write a "cmvn config" file (this is blank in the actual kaldi code, but it needs the argument passed)
+        cmvn_config = os.path.join(destination, 'online_cmvn.conf')
+        with open(cmvn_config, 'w', newline='') as cconf:
+            cconf.write("")
+
+        # Write a "splice config" file
+        splice_config = os.path.join(destination, 'splice.conf')
+        with open(splice_config, 'w', newline='') as sconf:
+            sconf.write('--left-context={}'.format(self.meta['splice_left_context']))
+            sconf.write('\n')
+            sconf.write('--right-context={}'.format(self.meta['splice_right_context']))
+
+        # Write a "config" file to input to the extraction binary
+        ext_config = os.path.join(destination, 'ivector_extractor.conf')
+        with open(ext_config, 'w', newline='') as ieconf:
+            ieconf.write('--cmvn-config={}\n'.format(cmvn_config))
+            ieconf.write('--ivector-period={}\n'.format(self.meta['ivector_period']))
+            ieconf.write('--splice-config={}\n'.format(splice_config))
+            ieconf.write('--lda-matrix={}\n'.format(os.path.join(destination, 'lda.mat')))
+            ieconf.write('--global-cmvn-stats={}\n'.format(os.path.join(destination, 'global_cmvn.stats')))
+            ieconf.write('--diag-ubm={}\n'.format(os.path.join(destination, 'final.dubm')))
+            ieconf.write('--ivector-extractor={}\n'.format(os.path.join(destination, 'final.ie')))
+            ieconf.write('--num-gselect={}\n'.format(self.meta['num_gselect']))
+            ieconf.write('--min-post={}\n'.format(self.meta['min_post']))
+            ieconf.write('--posterior-scale={}\n'.format(self.meta['posterior_scale']))
+            ieconf.write('--max-remembered-frames=1000\n')
+            ieconf.write('--max-count={}\n'.format(0))
+        return ext_config
